+++ CHANGES.rst
@@ -7,6 +7,9 @@ Changelog
 
 - Added endpoint to get booking schema.
   [daniele]
+- Added field "cosa_serve" (#40445).
+  [daniele]
+
 
 1.7.1 (2023-03-28)
 ------------------